Update README.md
Welcome to the Real Estate Price Prediction project! This project is aimed at predicting the prices of real estate properties in Pune and Bangalore through a web application. With the backend developed in Python and utilizing a MySQL database, this application provides insights into the potential prices of properties in these cities.

Features
Predicts the price of real estate properties in Pune and Bangalore.
Utilizes Python libraries such as NumPy, math, and matplotlib for data analysis and visualization.
Technologies Used
Python
MySQL
NumPy
math
matplotlib